refactor: replace input into ui folder
