+++ src/components/MoviePatchList/MoviePatchList.tsx
@@ -2,10 +2,10 @@ import { useTranslation } from 'next-i18next'
 import { FC } from 'react'
 
 import { IMovie } from '@/types/Movie'
+import { Input } from '@/ui/Input'
 
 import { MovieItem } from './MovieItem'
 import styles from './MoviePatchList.module.scss'
-import { SearchMovie } from './searchMovie'
 import { useMoviePatchList } from './useMoviePatchList'
 
 export interface MoviePatchListProps {
@@ -18,7 +18,12 @@ export const MoviePatchList: FC<MoviePatchListProps> = ({ initialMovies }) => {
 
   return (
     <div className={styles.moviePatchList}>
-      <SearchMovie text={searchText} setText={setSearchText} />
+      <Input
+        description={`${t('adminPage:findMovie')}`}
+        text={searchText}
+        setText={setSearchText}
+        placeholder={`${t('adminPage:find')}...`}
+      />
       {movies.length === 0 ? (
         <p>{t('adminPage:notFound')}</p>
       ) : (
