+++ src/ui/Input/Input.tsx
@@ -3,15 +3,22 @@ import { ChangeEventHandler, FC } from 'react'
 import styles from './Input.module.scss'
 
 export interface InputProps {
-  description: string
+  description?: string
+  placeholder?: string
   text: string
   setText: (str: string) => void
 }
 
-export const Input: FC<InputProps> = ({ description, text, setText }) => {
+export const Input: FC<InputProps> = ({
+  description = '',
+  placeholder = '',
+  text,
+  setText,
+}) => {
   const inputChangeHandler: ChangeEventHandler<HTMLInputElement> = (event) => {
     setText(event.target.value)
   }
+
   return (
     <label className={styles.label}>
       <span>{description}:</span>
@@ -20,6 +27,7 @@ export const Input: FC<InputProps> = ({ description, text, setText }) => {
         className={styles.input}
         value={text}
         onChange={inputChangeHandler}
+        placeholder={placeholder}
       />
     </label>
   )
